+++ unsync/unsync/commands/dbapi/__init__.py
@@ -1,0 +1,6 @@
+unsync_prefix = 'dbapi'
+
+unsync_commands = [
+    'commands.import',
+    'commands.export',
+]
