+++ unsync_pytds/unsync_pytds/__init__.py
@@ -1,0 +1,5 @@
+unsync_prefix = 'pytds'
+
+unsync_commands = [
+    'commands.connection',
+]
